+++ src/app/(searches)/shoppingList/page.jsx
@@ -1,8 +1,8 @@
 export default function Home() {
   return (
-    <div className="mx-3 mt-3">
-      <div className="flex text-center">
-        <div className="w-116 font-sans">買い物リスト</div>
+    <div className="sm:mx-3 sm:mt-3">
+      <div className="flex justify-center">
+        <div className=" font-sans">買い物リスト</div>
       </div>
     </div>
   );
